refactor(data-plane): decouple persisted formats from host pointer width
  Persisted capture and index data must remain portable across wasm32,
  wasm64, and native hosts instead of inheriting resident address-space
  limits. Keep cross-boundary coordinates fixed-width and reject values
  that cannot be represented by a current in-memory view.

+++ crates/logic_analyzer_graph_compiler/src/graph.rs
@@ -883,7 +883,7 @@ fn discover_live_capture_feature_from(
                 {
                     Some("live capture trigger channel references an unknown viewer channel")
                 } else if feature.session_plan().is_some_and(|plan| {
-                    plan.channel_count != feature.channels().len()
+                    plan.channel_count != feature.channels().len() as u64
                         || plan.sample_rate_hz as f64 != feature.sample_rate_hz()
                 }) {
                     Some("live capture session plan differs from its active channel/rate tuple")
